Finished CRUD operations on FE
The buttons for News and Posts from the navigation bar work properly now, the user avatar is clickable and display a menu, clicking "My Posts" will take you to the page displaying list added by the current user (which is hard coded by now), each card has a menu with Update and Delete options. The add and update forms are working properly.